PlotGenomeWideAlteration()

Function to graph genome-wide alteration based on cytoband level alteration table generated by CytobandLevelAlteration()

Usage

PlotGenomeWideAlteration(
  Cytoband.level.data,
  Genome = c("hg38", "hg19"),
  Coverage = 0.8
)

Arguments

Cytoband.level.data

Data frame of cytoband alterations

Genome

Whether samples are aligned to hg38 or hg19

Coverage

Threshold for coverage of cytobands needed to be included in the graph (default = 0.8)

Value

A list of graphs showing the frequency of each alteration in each cytoband across the cohort